Refactor broker to subclass ZmqJsonRpcTask
Broker handles RPC requests by first checking to see if the broker itself has a
handler for the command from the request.  If the broker has the relevant
handler, it is called and the result is returned through the router socket.  If
the broker does not have the relevant handler, handling is delegated to the
remote handler, if available.