Add serialization kwarg to `store` RPC method
The `serialization` kwarg can be set to one of:

 * `SERIALIZE__NONE` (default)
 * `SERIALIZE__JSON`
 * `SERIALIZE__PICKLE`

This value is added as a frame to data update messages published by the master,
as a result of calls to `store` RPC method.  The format of these messages is
now:

    [worker uuid][task uuid]['store'][serialization method][key][value]

where all frames are strings.  The intention here is that subscribers can use
the serialization method value to automatically de-serialize the value on
access.